replace flash.h and flash.c with flashF4.h and flashF4.c
